requirements.txt
fastapi
uvicorn
pydantic
pycoingecko
PyGithub
openai
requests
numpy
dotenv